+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Security vulerabilities in package dependencies are patched and updated.
+- axios proxy for mx-adapter
 
 ## [1.12.0]